refactor: extract shared test fixture into tests/helpers/mod.rs
Move TempDatabase and seed_project_sidebar_fixture out of
project_sidebar_filtering.rs into a shared helpers module so future
integration tests can reuse the fixture without copying SQL.

The local seed() method (multi-project counting fixture) stays in
project_sidebar_filtering.rs via a local impl block since it is specific
to that file's ordering and count tests.
