+++ tests/project_sidebar_filtering.rs
@@ -1,38 +1,17 @@
-use rusqlite::Connection;
-use std::path::PathBuf;
-use std::time::{SystemTime, UNIX_EPOCH};
+mod helpers;
 
-use sessions_chronicle::database::schema::initialize_database;
+use helpers::TempDatabase;
 use sessions_chronicle::database::{
     count_all_sessions, count_unassigned_sessions, has_unassigned_sessions, load_projects,
     load_sessions, load_sessions_for_filter, search_sessions, search_sessions_for_filter,
 };
 use sessions_chronicle::models::{AiAssistant, ProjectFilter};
 
-struct TempDatabase {
-    path: PathBuf,
-    connection: Connection,
-}
-
 impl TempDatabase {
-    fn new() -> Self {
-        let mut path = std::env::temp_dir();
-        let nanos = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_nanos();
-        path.push(format!(
-            "sessions-chronicle-project-sidebar-test-{}-{}.db",
-            std::process::id(),
-            nanos
-        ));
-
-        let connection = Connection::open(&path).expect("Failed to open temp database");
-        initialize_database(&connection).expect("Failed to initialize database");
-
-        Self { path, connection }
-    }
-
+    /// Seed the broader multi-project fixture used by count and ordering tests.
+    ///
+    /// Projects: alpha (id=1), beta (id=2), gamma (id=3)
+    /// Sessions: claude-alpha-1, claude-alpha-2 (project alpha), claude-unassigned-1, opencode-beta-1 (project beta)
     fn seed(&self) {
         self.connection
             .execute(
@@ -123,148 +102,11 @@ impl TempDatabase {
             )
             .expect("Failed to insert opencode beta session");
     }
-
-    fn seed_project_sidebar_fixture(&self) {
-        self.connection
-            .execute(
-                "INSERT INTO projects (id, path, name) VALUES (?1, ?2, ?3)",
-                rusqlite::params![1_i64, "/projects/alpha", "alpha"],
-            )
-            .expect("Failed to insert project alpha");
-
-        self.connection
-            .execute(
-                "INSERT INTO projects (id, path, name) VALUES (?1, ?2, ?3)",
-                rusqlite::params![2_i64, "/projects/beta", "beta"],
-            )
-            .expect("Failed to insert project beta");
-
-        self.connection
-            .execute(
-                "INSERT INTO sessions (id, tool, project_path, project_id, start_time, message_count, file_path, last_updated)
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8)",
-                rusqlite::params![
-                    "alpha-claude-old",
-                    "claude_code",
-                    Some("/projects/alpha"),
-                    Some(1_i64),
-                    10_i64,
-                    2_i64,
-                    "/tmp/alpha-claude-old.jsonl",
-                    100_i64,
-                ],
-            )
-            .expect("Failed to insert alpha old claude session");
-
-        self.connection
-            .execute(
-                "INSERT INTO sessions (id, tool, project_path, project_id, start_time, message_count, file_path, last_updated)
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8)",
-                rusqlite::params![
-                    "alpha-claude-new",
-                    "claude_code",
-                    Some("/projects/alpha"),
-                    Some(1_i64),
-                    20_i64,
-                    3_i64,
-                    "/tmp/alpha-claude-new.jsonl",
-                    200_i64,
-                ],
-            )
-            .expect("Failed to insert alpha new claude session");
-
-        self.connection
-            .execute(
-                "INSERT INTO sessions (id, tool, project_path, project_id, start_time, message_count, file_path, last_updated)
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8)",
-                rusqlite::params![
-                    "alpha-opencode",
-                    "opencode",
-                    Some("/projects/alpha"),
-                    Some(1_i64),
-                    30_i64,
-                    2_i64,
-                    "/tmp/alpha-opencode.jsonl",
-                    300_i64,
-                ],
-            )
-            .expect("Failed to insert alpha opencode session");
-
-        self.connection
-            .execute(
-                "INSERT INTO sessions (id, tool, project_path, project_id, start_time, message_count, file_path, last_updated)
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8)",
-                rusqlite::params![
-                    "unassigned-claude",
-                    "claude_code",
-                    Option::<String>::None,
-                    Option::<i64>::None,
-                    40_i64,
-                    1_i64,
-                    "/tmp/unassigned-claude.jsonl",
-                    400_i64,
-                ],
-            )
-            .expect("Failed to insert unassigned claude session");
-
-        self.connection
-            .execute(
-                "INSERT INTO sessions (id, tool, project_path, project_id, start_time, message_count, file_path, last_updated)
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8)",
-                rusqlite::params![
-                    "beta-claude",
-                    "claude_code",
-                    Some("/projects/beta"),
-                    Some(2_i64),
-                    50_i64,
-                    1_i64,
-                    "/tmp/beta-claude.jsonl",
-                    500_i64,
-                ],
-            )
-            .expect("Failed to insert beta claude session");
-
-        self.connection
-            .execute(
-                "INSERT INTO messages (session_id, message_index, role, content, timestamp, model)
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-                rusqlite::params![
-                    "unassigned-claude",
-                    0_i64,
-                    "user",
-                    "this session is lonely",
-                    1_i64,
-                    Option::<String>::None,
-                ],
-            )
-            .expect("Failed to insert message for unassigned claude session");
-
-        self.connection
-            .execute(
-                "INSERT INTO messages (session_id, message_index, role, content, timestamp, model)
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-                rusqlite::params![
-                    "alpha-claude-new",
-                    0_i64,
-                    "user",
-                    "alpha topic",
-                    2_i64,
-                    Option::<String>::None,
-                ],
-            )
-            .expect("Failed to insert message for alpha claude session");
-    }
-}
-
-impl Drop for TempDatabase {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_file(&self.path);
-    }
 }
 
 #[test]
 fn load_projects_orders_by_activity_and_keeps_zero_count_rows() {
-    let db = TempDatabase::new();
+    let db = TempDatabase::new("project-sidebar");
     db.seed();
 
     db.connection
@@ -330,7 +172,7 @@ fn load_projects_orders_by_activity_and_keeps_zero_count_rows() {
 
 #[test]
 fn project_sidebar_counts_include_unassigned_visibility_flag() {
-    let db = TempDatabase::new();
+    let db = TempDatabase::new("project-sidebar");
     db.seed();
 
     let all_count = count_all_sessions(&db.path, &[AiAssistant::ClaudeCode])
@@ -363,7 +205,7 @@ fn project_sidebar_counts_include_unassigned_visibility_flag() {
 
 #[test]
 fn load_sessions_for_filter_returns_project_and_tool_intersection() {
-    let db = TempDatabase::new();
+    let db = TempDatabase::new("project-sidebar");
     db.seed_project_sidebar_fixture();
 
     let sessions = load_sessions_for_filter(
@@ -379,7 +221,7 @@ fn load_sessions_for_filter_returns_project_and_tool_intersection() {
 
 #[test]
 fn search_sessions_for_filter_returns_only_unassigned_matches() {
-    let db = TempDatabase::new();
+    let db = TempDatabase::new("project-sidebar");
     db.seed_project_sidebar_fixture();
 
     let sessions = search_sessions_for_filter(
@@ -397,7 +239,7 @@ fn search_sessions_for_filter_returns_only_unassigned_matches() {
 
 #[test]
 fn load_sessions_wrapper_matches_all_sessions_filter() {
-    let db = TempDatabase::new();
+    let db = TempDatabase::new("project-sidebar");
     db.seed_project_sidebar_fixture();
 
     let tools = &[AiAssistant::ClaudeCode];
@@ -419,7 +261,7 @@ fn load_sessions_wrapper_matches_all_sessions_filter() {
 
 #[test]
 fn search_sessions_wrapper_matches_all_sessions_filter() {
-    let db = TempDatabase::new();
+    let db = TempDatabase::new("project-sidebar");
     db.seed_project_sidebar_fixture();
 
     let tools = &[AiAssistant::ClaudeCode];
